requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
PyMySQL==0.6.3
argparse==1.2.1
itsdangerous==0.24
wsgiref==0.1.2
boto==2.28.0